move hooks around
