+++ javascript/react-controls/src/index.ts
@@ -3,5 +3,4 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 export * from './controls/index.js'
-export * from './hooks.js'
 export * from './styles.js'
