+++ javascript/react-hooks/package.json
@@ -32,7 +32,6 @@
 	},
 	"dependencies": {
 		"@data-wrangling-components/core": "workspace:^",
-		"@data-wrangling-components/react-controls": "workspace:^",
 		"@essex/arquero": "^2.0.2",
 		"@essex/dataflow": "^0.1.1"
 	},
